refactor: сhange structure
