+++ app/components/errorBoundary/index.tsx
@@ -1,6 +1,7 @@
 import type { Route } from '../../+types/root';
 import { isRouteErrorResponse } from 'react-router';
 import styles from './ErrorBoundary.module.css';
+// import { messages } from '@/sources/messages';
 
 export function ErrorBoundaryComponent({
   error,
