--- app/pages/home-page/HomePage.tsx
@@ -1,11 +1,0 @@
-import styles from './HomePage.module.css';
-
-export const HomePage = () => {
-  console.log('Hello');
-
-  return (
-    <div className={styles.container}>
-      <h2>Home Page</h2>
-    </div>
-  );
-};
