+++ app/routes.ts
@@ -1,7 +1,8 @@
 import { type RouteConfig, index, route } from '@react-router/dev/routes';
+import { AppRoutes } from './sources/enums';
 
 export default [
-  index('routes/home.tsx'),
-  route('about', 'routes/about.tsx'),
-  route('*', 'routes/404.tsx'),
+  index('routes/mainPage/index.tsx'),
+  route(AppRoutes.ABOUT, 'routes/aboutPage/index.tsx'),
+  route(AppRoutes.NOT_FOUND, 'routes/notFoundPage/index.tsx'),
 ] satisfies RouteConfig;
