+++ app/sources/enums.ts
@@ -2,3 +2,9 @@ export enum Variant {
   PRIMARY = 'primary',
   SECONDARY = 'secondary',
 }
+
+export enum AppRoutes {
+  HOME = '/',
+  ABOUT = 'about',
+  NOT_FOUND = '*',
+}
